+++ open-sse/providers/registry/index.js
@@ -133,6 +133,7 @@ import p131 from "./zenmux-free.js";
 import p132 from "./aihubmix.js";
 import p133 from "./aipass.js";
 import p134 from "./tokenharbor.js";
+import p135 from "./nara.js";
 
 export default [
   p0,
@@ -269,4 +270,5 @@ export default [
   p132,
   p133,
   p134,
+  p135,
 ];